readeasy-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# Authentication
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.1.3

# Database
redis[hiredis]>=5.0.1

# Document processing
PyPDF2>=3.0.0
Pillow>=10.1.0
python-magic

# AI services
mistralai>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.1
